apps/api/backend/requirements.txt
fastapi
uvicorn
httpx
aiofiles
python-multipart
openai
python-dotenv
supabase
moviepy
imageio[ffmpeg]
python-dotenv
librosa
soundfile
numpy
scipy
replicate

